+++ CHANGELOG.md
@@ -45,11 +45,14 @@
 * Icon themes: 'scaled' directories are now ignored ([#459][459]).
 * Border radius limited by max(horizontal-pad, vertical-pad), instead
   of min(horizontal-pad, vertical-pad) ([#436][436]).
+* `stdout` and `stderr` of the launched application is now preserved,
+  instead of being redirected to `/dev/null` ([#480][480]).
 
 [419]: https://codeberg.org/dnkl/fuzzel/issues/419
 [439]: https://codeberg.org/dnkl/fuzzel/issues/439
 [459]: https://codeberg.org/dnkl/fuzzel/issues/459
 [436]: https://codeberg.org/dnkl/fuzzel/issues/436
+[480]: https://codeberg.org/dnkl/fuzzel/issues/480
 
 
 ### Deprecated